requirements.txt
aioredis==1.3.1
amqp==5.0.6
appdirs==1.4.3
APScheduler==3.6.3
apturl==0.5.2
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
Authlib==0.15.3
autobahn==21.1.1
Automat==20.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
blinker==1.4
Brlapi==0.7.0
bs4==0.0.1
CacheControl==0.12.6
cachetools==4.2.2
celery==5.0.5
celery-progress==0.1.0
certifi==2019.11.28
cffi==1.14.5
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
Click==7.0
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
command-not-found==0.3
constantly==15.1.0
contextlib2==0.6.0
cryptography==2.8
cupshelpers==1.0
daphne==2.5.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==2.2.24
django-3-jet==1.0.8
django-ckeditor==5.8.0
django-cors-headers==3.1.0
django-crontab==0.7.1
django-extensions==3.1.0
django-froala-editor==3.2.2
django-js-asset==1.2.2
django-limits==0.0.6
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-restframework==0.0.1
django-saml2-auth==2.2.1
django-user-agents==0.4.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
elementpath==2.2.1
entrypoints==0.3
et-xmlfile==1.0.1
Faker==8.0.0
filelock==3.0.12
ftpretty==0.3.4
gunicorn==20.0.4
hiredis==1.1.0
html5lib==1.0.1
httplib2==0.14.0
hyperlink==21.0.0
idna==2.8
imageio==2.9.0
imageio-ffmpeg==0.4.2
importlib-metadata==1.5.0
importlib-resources==5.1.2
incremental==17.5.0
ipaddr==2.2.0
jdcal==1.4.1
keyring==18.0.1
kombu==5.0.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
more-itertools==4.2.0
moviepy==1.0.3
msgpack==0.6.2
Naked==0.1.31
netifaces==0.10.4
nltk==3.4.5
numpy==1.19.4
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.5
packaging==20.3
pbr==5.5.1
pdfkit==0.6.1
pep517==0.8.2
pexpect==4.6.0
Pillow==7.0.0
proglog==0.1.9
progress==1.5
prompt-toolkit==3.0.18
protobuf==3.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.10.1
pycups==1.9.73
PyGObject==3.36.0
pygraphviz==1.7
PyHamcrest==2.0.2
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==20.0.1
pyparsing==2.4.6
pyRFC3339==1.1
pysaml2==6.5.1
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-telegram-bot==13.6
pytoml==0.1.21
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
retrying==1.3.3
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
selenium==3.141.0
service-identity==18.1.0
shellescape==3.8.1
simple-salesforce==1.11.1
simplejson==3.16.0
six==1.14.0
soupsieve==2.0.1
sqlparse==0.4.1
systemd-python==234
testresources==2.0.1
text-unidecode==1.3
toml==0.10.2
tornado==6.1
tqdm==4.54.1
twilio==6.60.0
Twisted==20.3.0
txaio==20.12.1
typing-extensions==3.7.4.3
tzlocal==2.1
ua-parser==0.8.0
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
user-agents==2.0
vine==5.0.0
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
-e git://github.com/paulosman/django-xframeoptions@4147688d094b2d57e46620b636c01e78219ffd8c#egg=xframeoptions
xkit==0.0.0
xlrd==2.0.1
xlwt==1.3.0
xmlschema==1.5.3
zipp==1.0.0
zope.interface==5.2.0
